fix: rearrange code to fix cryptic build system error
It did _not_ like the second `await` at the bottom of `ExternalListener.BindAsync()`. It kept giving errors about how the type `Random` (which isn't used in the code at all) is from an unreferenced assembly.

I spent three hours trying to figure this out because, as you can imagine, the error was 100% useless and a total red herring. I even uninstalled and reinstalled Visual Studio and Unity. It wasn't until, on a lark, I decided to try reverting to the previous commit, and the error went away. It was then a straightforward process of finding which exact change broke things.

I assume it had something to do with the desugaring of the `async` method interacting poorly with the build system but I'm still not entirely sure. Either way, that's three hours of my life I will never get back.