Add support for Unity Relay (#8)
* WIP implementing Unity Relay integration

* more Unity Relay integration work

* set up an external connection per peer

* check-in `SockAddr`

* integrate `ExternalAddressBookEntry`

* a couple fixes from testing

* fix errors after rebase

* get TNT working (sort of)

* fix: update the unit tests for the recent changes

* fix: match the native layout for SockAddr

The AddressFamily enum is 4 bytes, but C uses 2 bytes.

I think the _dataSpan field added some bytes which we want to avoid.

The constructor taking an IPEndPoint always wrote the address to
_data[2..], but that's not correct for IPv6.

* got it as far as requesting the first transmit, getting SIGSEGV

* fix: prevent warnigns about unused _ for exception names

* chore: update the declarations of tce_external_transmit_get_addr and get_packet

* fix: AddressFamily's values don't match the AF_* definitions

* chore: add ToString for SockAddr

* fix: implement the disposable pattern correctly

* fix: make sure the networkDriver and the packet buffer are disposed

* fix: use correct AddressFamily constants per platform

* fix: rearrange code to fix cryptic build system error

It did _not_ like the second `await` at the bottom of `ExternalListener.BindAsync()`. It kept giving errors about how the type `Random` (which isn't used in the code at all) is from an unreferenced assembly.

I spent three hours trying to figure this out because, as you can imagine, the error was 100% useless and a total red herring. I even uninstalled and reinstalled Visual Studio and Unity. It wasn't until, on a lark, I decided to try reverting to the previous commit, and the error went away. It was then a straightforward process of finding which exact change broke things.

I assume it had something to do with the desugaring of the `async` method interacting poorly with the build system but I'm still not entirely sure. Either way, that's three hours of my life I will never get back.

* fix: prevent leaking NetworkDriver and NetworkConnection

* fix: enable TashiNetworkDriver to be stopped and started

* fix: set the total number of connections when allocating a relay

* chore: add a property to indicate whether the session has started

* fix: implement Equals and GetHashCode for SockAddr

* fix: ensure we always call NetworkDriver.EndSend

* fix: tce_external_recv_prepare's bufLen is now a UInt64

* fix: don't use IntPtr for 64 bit lengths

* fix: don't overwrite the clientId when sending over Relay

* chore: stop using the lobby once the session has begun

* chore: print more details when an address book entry is added

* feat: add `SockAddrTests`, fix a bunch of bugs

* attempted fixes for unreliability

* chore: connections are always disconnected on creation

* chore: add some TODOs for connect/disconnect

* fix: call ScheduleUpdate once per frame per driver

* fix: include the client id length when creating a writer with BeginSend

* fix: don't log that pending connections are unknown

* chore: use NetworkConnection's State property

* fix: enable receiving before the platform has started

This is required because some nodes will begin sending data before other nodes
know the session is ready to start.

* chore: remove incorrect TODOs

InvokeOnTransportEvent will only be called in response to events that
have reached consensus.

* chore: improve the example's status text

* chore: remove /.idea

* chore: code review fixes

* chore: the next version will be 0.2.0

* fix: the link for the first release was broken

* chore: update CHANGELOG with 0.1.0's limitations

* chore: include the full CHANGELOG.md in the GitHub release

* fix: refactor external transmits to mitigate potential memory leak

This also eliminates the redundant copy to an intermediate buffer.

---------

Co-authored-by: Austin Bonander <austin@launchbadge.com>