fix: work around an error relating to UnityEngine.Random
A developer reported a bug about a missing assembly reference to
Unity.Mathematics.Random when using Build and Run on macOS.

I couldn't reproduce it on my Mac and the error doesn't really make
sense, because we're using UnityEngine.Random. It's possible that
UnityEngine.Random depends on Unity.Mathematics.Random, or that the
developer's environment has some issues, but either way we don't need to
randomize the color of the sprice in the example project.